Adding 5 more questions and score color threshold
